+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Changed
+
 - Updates __init__ signature kwargs replaced with parent for better documentation.
+
+- Update task cancel function defintions. Add success, failure cancel tests
 
 ## [0.35.0] - 2023-05-02